Move config to platform insets
