+++ compose/ui/ui/src/notMobileMain/kotlin/androidx/compose/ui/platform/PlatformInsets.notMobile.kt
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.window
+package androidx.compose.ui.platform
 
-internal actual var CurrentRootLayoutInsetsConfig: RootLayoutInsetsConfig =
-    DefaultRootLayoutInsetsConfig
+internal actual var PlatformInsetsConfig: InsetsConfig = ZeroInsetsConfig
