+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/window/Dialog.skiko.kt
@@ -34,8 +34,10 @@ import androidx.compose.ui.input.pointer.PointerButton
 import androidx.compose.ui.input.pointer.PointerEventType
 import androidx.compose.ui.input.pointer.PointerInputEvent
 import androidx.compose.ui.layout.Layout
-import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.platform.InsetsConfig
 import androidx.compose.ui.platform.PlatformInsets
+import androidx.compose.ui.platform.PlatformInsetsConfig
+import androidx.compose.ui.platform.ZeroInsetsConfig
 import androidx.compose.ui.requireCurrent
 import androidx.compose.ui.semantics.dialog
 import androidx.compose.ui.semantics.semantics
@@ -177,7 +179,7 @@ private fun DialogLayout(
     onOutsidePointerEvent: ((PointerInputEvent) -> Unit)? = null,
     content: @Composable () -> Unit
 ) {
-    val platformInsets = properties.insetsConfig.insets
+    val platformInsets = properties.insetsConfig.safeInsets
     RootLayout(
         modifier = modifier,
         focusable = true,
@@ -189,7 +191,7 @@ private fun DialogLayout(
         ) {
             owner.bounds = it
         }
-        properties.insetsConfig.excludeInsets {
+        properties.insetsConfig.excludeSafeInsets {
             Layout(
                 content = content,
                 measurePolicy = measurePolicy
@@ -198,8 +200,8 @@ private fun DialogLayout(
     }
 }
 
-private val DialogProperties.insetsConfig: RootLayoutInsetsConfig
-    get() = if (usePlatformInsets) CurrentRootLayoutInsetsConfig else DefaultRootLayoutInsetsConfig
+private val DialogProperties.insetsConfig: InsetsConfig
+    get() = if (usePlatformInsets) PlatformInsetsConfig else ZeroInsetsConfig
 
 @Composable
 private fun rememberDialogMeasurePolicy(
