+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/window/RootLayout.skiko.kt
@@ -158,26 +158,6 @@ internal fun MeasureScope.positionWithInsets(
     return position + offset
 }
 
-internal interface RootLayoutInsetsConfig {
-    val insets: PlatformInsets
-        @Composable get
-
-    @Composable
-    fun excludeInsets(content: @Composable () -> Unit)
-}
-
-internal object DefaultRootLayoutInsetsConfig : RootLayoutInsetsConfig {
-    override val insets: PlatformInsets
-        @Composable get() = PlatformInsets.Zero
-
-    @Composable
-    override fun excludeInsets(content: @Composable () -> Unit) {
-        content()
-    }
-}
-
-internal expect var CurrentRootLayoutInsetsConfig: RootLayoutInsetsConfig
-
 private fun Density.platformDefaultConstrains(
     constraints: Constraints
 ): Constraints = constraints.copy(
