--- compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/window/RootLayout.uikit.kt
@@ -1,45 +1,0 @@
-/*
- * Copyright 2023 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package androidx.compose.ui.window
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.InternalComposeApi
-import androidx.compose.ui.platform.LocalLayoutMargins
-import androidx.compose.ui.platform.LocalSafeArea
-import androidx.compose.ui.platform.PlatformInsets
-import androidx.compose.ui.platform.exclude
-
-@OptIn(InternalComposeApi::class)
-private object SafeAreaRootLayoutInsetsConfig : RootLayoutInsetsConfig {
-    override val insets: PlatformInsets
-        @Composable get() = LocalSafeArea.current
-
-    @Composable
-    override fun excludeInsets(content: @Composable () -> Unit) {
-        val safeArea = LocalSafeArea.current
-        val layoutMargins = LocalLayoutMargins.current
-        CompositionLocalProvider(
-            LocalSafeArea provides PlatformInsets(),
-            LocalLayoutMargins provides layoutMargins.exclude(safeArea),
-            content = content
-        )
-    }
-}
-
-internal actual var CurrentRootLayoutInsetsConfig: RootLayoutInsetsConfig =
-    SafeAreaRootLayoutInsetsConfig
